+++ src/vue/views/App.vue
@@ -1,20 +1,21 @@
 <style lang="less">
-  /* Global */
+  /* Semantic UI Global */
   & { @import "~semantic-ui-less/definitions/globals/reset"; }
   & { @import "~semantic-ui-less/definitions/globals/site"; }
-  /* Elements */
+  /* Semantic UI Elements */
   & { @import "~semantic-ui-less/definitions/elements/button"; }
   & { @import "~semantic-ui-less/definitions/elements/container"; }
+  & { @import "~semantic-ui-less/definitions/elements/divider"; }
   & { @import "~semantic-ui-less/definitions/elements/header"; }
   & { @import "~semantic-ui-less/definitions/elements/icon"; }
   & { @import "~semantic-ui-less/definitions/elements/image"; }
   & { @import "~semantic-ui-less/definitions/elements/label"; }
   & { @import "~semantic-ui-less/definitions/elements/list"; }
   & { @import "~semantic-ui-less/definitions/elements/segment"; }
-  /* Collections */
+  /* Semantic UI Collections */
   & { @import "~semantic-ui-less/definitions/collections/grid"; }
   & { @import "~semantic-ui-less/definitions/collections/menu"; }
-  /* Views */
+  /* Semantic UI Views */
   & { @import "~semantic-ui-less/definitions/views/card"; }
   & { @import "~semantic-ui-less/definitions/views/comment"; }
   & { @import "~semantic-ui-less/definitions/views/item"; }